chore: rename core components
