+++ app/(dashboard)/layout.tsx
@@ -1,7 +1,7 @@
 import NavProgress from '@/components/NavProgress';
-import { MainNav } from '@/components/core/MainNav';
-import ThemeToggle from '@/components/core/ThemeToggle';
-import { UserNav } from '@/components/core/UserNav';
+import MainNav from '@/components/core/main-nav';
+import ThemeToggle from '@/components/core/theme-toggle';
+import UserNav from '@/components/core/user-nav';
 import { getCurrentUser } from '@/utils/auth';
 import { Flex } from '@tremor/react';
 import { Metadata } from 'next';
